Renamed purple arrow images
