+++ bloom-and-vine/src/Images/purpleForwardArrow.png
